Add confirmation to destructive operations.

+++ migration.py
@@ -8,17 +8,16 @@ from tfc_migrate import \
             agent_pools, workspace_vars, run_triggers, state_versions, \
                 policy_set_params
 
-# TODO: review imports like urllib / ast
-# TODO: convert all string formating to f format
-# TODO: make sure the source api has all of the resources that we migrate.
+# TODO: ast import?
+# TODO: determine which of the unused functions we can delete, otherwise implement them properly
+# TODO: the maps that are output need to be more explicit and not just rely on key-value
+# TODO: similarly, clean up TFE_VCS_CONNECTION_MAP
+# TODO: remove the TFE_* from all migrate function calls, retrieve that information from the API object itself.
+# TODO: Create a TF config file that will create all the resources that are needed to test a migration
 # TODO: use a logger instead of print statements
-# TODO: have each module handle it's own output writes
 # TODO: confirm all functions are idempotent
-# TODO: try not to repeat the try/except blocks
 # TODO: logging should have tabs to be more readable
 # TODO: create a base migrate class w/ a logger
-# TODO: clean up TFE_VCS_CONNECTION_MAP
-# TODO: remove the TFE_* from all migrate functino calls, retrieve that information from the API object itself.
 
 
 # Source Org
@@ -32,57 +31,44 @@ TFE_URL_TARGET = os.getenv("TFE_URL_TARGET", None)
 TFE_ORG_TARGET = os.getenv("TFE_ORG_TARGET", None)
 TFE_VCS_CONNECTION_MAP = ast.literal_eval(os.getenv("TFE_VCS_CONNECTION_MAP", None))
 
+def confirm_delete(target_resource_type_str):
+    answer = ""
+    while answer not in ["y", "n"]:
+        # TODO: add target url and org
+        question_string = \
+            "This will destroy all %s. Are you sure you want to continue? [Y/N]: " % target_resource_type_str
+        answer = input(question_string).lower()
+    return answer == "y"
 
-def print_output(\
+def handle_output(\
     teams_map, ssh_keys_map, ssh_key_name_map, workspaces_map, \
         workspace_to_ssh_key_map, workspace_to_config_version_map, \
             policies_map, policy_sets_map, sensitive_policy_set_parameter_data, \
-                sensitive_variable_data):
-    lines = [
-        "MIGRATION MAPS:",
-        "teams_map: %s\n\n" % teams_map,
-        "ssh_keys_map: %s\n\n" % ssh_keys_map,
-        "ssh_key_name_map: %s\n\n" % ssh_key_name_map,
-        "workspaces_map: %s\n\n" % workspaces_map,
-        "workspace_to_ssh_key_map: %s\n\n" % workspace_to_ssh_key_map,
-        "workspace_to_config_version_map: %s\n\n" % workspace_to_config_version_map,
-        "policies_map: %s\n\n" % policies_map,
-        "policy_sets_map: %s\n\n" % policy_sets_map,
-        "sensitive_policy_set_parameter_data: %s\n\n" % sensitive_policy_set_parameter_data,
-        "sensitive_variable_data: %s\n\n" % sensitive_variable_data
-    ]
+                sensitive_variable_data, write_to_file=False):
 
-    for line in lines:
-        print(line)
+        output_json = {
+            "teams_map": teams_map,
+            "ssh_keys_map": ssh_keys_map,
+            "ssh_key_name_map": ssh_key_name_map,
+            "workspaces_map": workspaces_map,
+            "workspace_to_ssh_key_map": workspace_to_ssh_key_map,
+            "workspace_to_config_version_map": workspace_to_config_version_map,
+            "policies_map": policies_map,
+            "policy_sets_map": policy_sets_map,
+            "sensitive_policy_set_parameter_data": sensitive_policy_set_parameter_data,
+            "sensitive_variable_data": sensitive_variable_data
+        }
 
-
-def write_output(\
-    teams_map, ssh_keys_map, ssh_key_name_map, workspaces_map, \
-        workspace_to_ssh_key_map, workspace_to_config_version_map, \
-            policies_map, policy_sets_map, sensitive_policy_set_parameter_data, \
-                sensitive_variable_data):
-
-        lines = [
-            "teams_map: %s\n\n" % teams_map,
-            "ssh_keys_map: %s\n\n" % ssh_keys_map,
-            "ssh_key_name_map: %s\n\n" % ssh_key_name_map,
-            "workspaces_map: %s\n\n" % workspaces_map,
-            "workspace_to_ssh_key_map: %s\n\n" % workspace_to_ssh_key_map,
-            "workspace_to_config_version_map: %s\n\n" % workspace_to_config_version_map,
-            "policies_map: %s\n\n" % policies_map,
-            "policy_sets_map: %s\n\n" % policy_sets_map,
-            "sensitive_policy_set_parameter_data: %s\n\n" % sensitive_policy_set_parameter_data,
-            "sensitive_variable_data: %s\n\n" % sensitive_variable_data
-        ]
-
-        with open("outputs.txt", "w") as f:
-            f.write("".join(lines))
+        if write_to_file:
+            with open("outputs.txt", "w") as f:
+                f.write(output_json)
+        else:
+            print(output_json)
 
 
 def migrate_to_target(api_source, api_target, write_to_file):
     teams_map = teams.migrate(api_source, api_target)
 
-    # TODO: use it or remove it
     # TODO: org_membership_map = org_memberships.migrate(api_source, api_target, teams_map)
 
     ssh_keys_map, ssh_key_name_map = ssh_keys.migrate_keys(api_source, api_target)
@@ -92,13 +78,11 @@ def migrate_to_target(api_source, api_target, write_to_file):
     workspaces_map, workspace_to_ssh_key_map = \
         workspaces.migrate(api_source, api_target, TFE_VCS_CONNECTION_MAP, agent_pools_map)
 
-    # TODO: use it or remove it
-    # ssh_keys.migrate_key_files(api_target, ssh_key_name_map, ssh_key_file_path_map)
-
+    # TODO: ssh_keys.migrate_key_files(api_target, ssh_key_name_map, ssh_key_file_path_map)
     workspaces_map, workspace_to_ssh_key_map = \
         workspaces.migrate(api_source, api_target, TFE_VCS_CONNECTION_MAP, agent_pools_map)
 
-    state_versions.migrate_current(api_source, api_target, TFE_ORG_SOURCE, workspaces_map)
+    state_versions.migrate_current(api_source, api_target, workspaces_map)
 
     """
     NOTE: if you wish to generate a map of Sensitive variables that can be used to update
@@ -107,13 +91,11 @@ def migrate_to_target(api_source, api_target, write_to_file):
     # TODO: right place for this note?
     """
     # TODO: is this var name accurate?
-    sensitive_variable_data = workspace_vars.migrate(
-        api_source, api_target, workspaces_map)
+    sensitive_variable_data = workspace_vars.migrate(api_source, api_target, workspaces_map)
 
-    # TODO: doesn't this happen in the normal migrate function now?
-    # workspace_vars.migrate_sensitive(api_target, sensitive_variable_data_map)
+    # TODO: workspace_vars.migrate_sensitive(api_target, sensitive_variable_data_map)
 
-    workspaces.migrate_ssh_keys(
+    workspaces.migrate_ssh_keys( \
         api_source, api_target, workspaces_map, workspace_to_ssh_key_map, ssh_keys_map)
 
     run_triggers.migrate(api_source, api_target, workspaces_map)
@@ -125,11 +107,10 @@ def migrate_to_target(api_source, api_target, write_to_file):
     workspace_to_config_version_map = config_versions.migrate( \
         api_source, api_target, workspaces_map)
 
-    # TODO: fix logging
-    # config_versions.migrate_config_files(\
+    # TODO: config_versions.migrate_config_files(\
     #   api_target, workspace_to_config_version_map, workspace_to_file_path_map)
 
-    policies_map = policies.migrate(api_source, api_target, TFE_TOKEN_SOURCE, TFE_URL_SOURCE)
+    policies_map = policies.migrate(api_source, api_target, TFE_TOKEN_SOURCE)
 
     policy_sets_map = policy_sets.migrate(\
         api_source, api_target, TFE_VCS_CONNECTION_MAP, workspaces_map, policies_map)
@@ -139,48 +120,48 @@ def migrate_to_target(api_source, api_target, write_to_file):
     sensitive_policy_set_parameter_data = \
         policy_set_params.migrate(api_source, api_target, policy_sets_map)
 
-    # TODO: what is this function
+    # TODO: what is this function?
     # policy_set_params.migrate_sensitive(api_target, sensitive_policy_set_parameter_data_map)
 
     registry_modules.migrate(api_source, api_target, TFE_VCS_CONNECTION_MAP)
 
-    if write_to_file:
-        write_output(teams_map, ssh_keys_map, ssh_key_name_map, workspaces_map, \
-                workspace_to_ssh_key_map, workspace_to_config_version_map, \
-                    policies_map, policy_sets_map, sensitive_policy_set_parameter_data, \
-                        sensitive_variable_data)
-    else:
-        print_output(teams_map, ssh_keys_map, ssh_key_name_map, workspaces_map, \
-                workspace_to_ssh_key_map, workspace_to_config_version_map, \
-                    policies_map, policy_sets_map, sensitive_policy_set_parameter_data, \
-                        sensitive_variable_data)
+    handle_output(teams_map, ssh_keys_map, ssh_key_name_map, workspaces_map, \
+            workspace_to_ssh_key_map, workspace_to_config_version_map, \
+                policies_map, policy_sets_map, sensitive_policy_set_parameter_data, \
+                    sensitive_variable_data, write_to_file=write_to_file)
 
 
-def delete_all_from_target(api):
-    # TODO: confirm all of these before running, since they are destructive
-
+def delete_all_from_target(api, no_confirmation):
     # Deleting workspaces allows us to skip deleting notification configs,
-    # config_versions, state_versions, workspace_vars.
-    workspaces.delete_all(api)
+    # config_versions, run_triggers, state_versions, workspace_vars.
+    if no_confirmation or confirm_delete("workspaces"):
+        workspaces.delete_all(api)
 
     # No need to delete the key files, they get deleted as well.
-    ssh_keys.delete_all_keys(api)
+    if no_confirmation or confirm_delete("SSH keys"):
+        ssh_keys.delete_all_keys(api)
 
-    teams.delete_all(api)
+    if no_confirmation or confirm_delete("teams"):
+        teams.delete_all(api)
 
     # Delete all the policy sets before deleting the individual policies.
     # No need to delete policy set params since we delete the entire set.
-    policy_sets.delete_all(api)
-    policies.delete_all(api)
+    if no_confirmation or confirm_delete("policy sets"):
+        policy_sets.delete_all(api)
 
-    registry_modules.delete_all(api)
+    if no_confirmation or confirm_delete("policies"):
+        policies.delete_all(api)
 
-    agent_pools.delete_all(api)
+    if no_confirmation or confirm_delete("registry modules"):
+        registry_modules.delete_all(api)
+
+    if no_confirmation or confirm_delete("agent pools"):
+        agent_pools.delete_all(api)
 
 
-def main(api_source, api_target, write_to_file, delete_all):
+def main(api_source, api_target, write_to_file, delete_all, no_confirmation):
     if delete_all:
-        delete_all_from_target(api_target)
+        delete_all_from_target(api_target, no_confirmation)
     else:
         migrate_to_target(api_source, api_target, write_to_file)
 
@@ -194,6 +175,7 @@ if __name__ == "__main__":
     parser = argparse.ArgumentParser(description='Migrate from TFE/C to TFE/C.')
     parser.add_argument('--write-output', dest="write_output", action="store_true", help="Write output to a file.")
     parser.add_argument('--delete-all', dest="delete_all", action="store_true", help="Delete all resources from the target API.")
+    parser.add_argument('--no-confirmation', dest="no_confirmation", action="store_true", help="If set, don't ask for confirmation before deleting all target resources.")
     args = parser.parse_args()
 
     api_source = TFC(TFE_TOKEN_SOURCE, url=TFE_URL_SOURCE)
@@ -202,4 +184,4 @@ if __name__ == "__main__":
     api_target = TFC(TFE_TOKEN_TARGET, url=TFE_URL_TARGET)
     api_target.set_org(TFE_ORG_TARGET)
 
-    main(api_source, api_target, args.write_output, args.delete_all)
+    main(api_source, api_target, args.write_output, args.delete_all, args.no_confirmation)
